requirements.txt
scikit-learn
numpy
openai
requests
psycopg2-binary
fastapi
uvicorn[standard]